+++ Makefile
@@ -1,0 +1,4 @@
+.PHONY: docs
+
+docs:
+		python docs/scripts/generate_pages.py
