+++ addons/flux-cronjob-deployment/README.md
@@ -1,0 +1,3 @@
+# Template for Deploying a Cronjob to Kubernetes with FluxCD
+
+A Template for deploying a cronjob to Kubernetes using FluxCD in the Titan environment.
