+++ addons/flux-service-deployment/README.md
@@ -1,0 +1,3 @@
+# Deploying a Service to Kubernetes with FluxCD
+
+A Template for deploying a service to Kubernetes with FluxCD in the Titan environment.
