+++ addons/github-workflows/README.md
@@ -1,1 +1,3 @@
 # Github Workflows
+
+Template for creating a Github workflows.
