+++ addons/mkdocs/README.md
@@ -1,3 +1,14 @@
-# MKDocs
+# MKDocs Template Documentation
 
-This template leverages MKDocs to generate static documentation sites from Markdown files. It utilizes the Material theme for a modern and responsive design. Additionally, the template is configured to deploy the generated static HTML site to GitHub Pages, making it easy to host and share your documentation online.
+A template for creating and deploying documentation using [MKDocs](https://www.mkdocs.org/), leveraging the [Material theme](https://squidfunk.github.io/mkdocs-material/) for a clean and modern look.  The included CI/CD pipeline facilitates seamless deployment to GitHub Pages.
+
+## Template Structure
+
+*   **`mkdocs.yaml`**: Configuration file for MKDocs, defining the site's appearance, navigation, and deployment settings.
+*   **`.github/workflows/`**: Contains the GitHub Actions workflow for automatic deployment.
+*   **`docs/`**: Contains the markdown files, configuration files, and static assets that make up the documentation.
+*   **`docs/.pages`:**  Important file containing the `nav` definition for the site's navigation.
+*   **`docs/requirements.txt`**: Lists the necessary Python packages.
+*   **`docs/static/`**:  Folder for static assets like logos and favicons.
+*   **`docs/index.md`**: Defaults to referencing the projects `README.md` file.
+*   **(Other files like `getting-started.md`, `faq.md` etc.):**  Markdown files containing the actual documentation content.
