+++ bases/sfmc-app/README.md
@@ -1,0 +1,3 @@
+# SFMC Application Template
+
+A template for deploying a Salesforce Marketing Cloud (SFMC) application using Docker and PM2.  It's designed to streamline the deployment process and provide a structured approach to interacting with Journey Builder activities.
