+++ docs/requirements.txt
@@ -1,5 +1,5 @@
-mkdocs==1.6.1
-mkdocs-material==9.5.34
-mkdocs-include-markdown-plugin==6.2.2
-mkdocs-awesome-pages-plugin==2.9.3
-pymdown-extensions==10.9
+mkdocs~=1.6.1
+mkdocs-material~=9.5.34
+mkdocs-include-markdown-plugin~=6.2.2
+mkdocs-awesome-pages-plugin~=2.9.3
+pymdown-extensions~=10.14
