+++ docs/template/usage.md
@@ -1,0 +1,17 @@
+## Usage Instructions¶
+
+The preferred way to use the template is through the [tam-cli](https://github.com/DND-IT/tam-cli) tool. This tool allows you to easily initialize a new project with the template.
+
+### Installation
+
+```shell
+pipx install git+https://github.com/DND-IT/tam-cli.git
+```
+
+### Running
+
+```shell
+tam template init [REPO_NAME]
+```
+
+Follow the prompts to select the template.
