+++ docs/templates_/.pages
@@ -1,0 +1,5 @@
+nav:
+  - index.md
+  - bases
+  - addons
+  - external
