+++ docs/templates_/addons/discovery-app-layer.md
@@ -1,0 +1,5 @@
+
+{% include "../../../addons/discovery-app-layer/README.md" %}
+
+{% include "template/usage.md" %}
+
