+++ docs/templates_/addons/flux-cronjob-deployment.md
@@ -1,0 +1,5 @@
+
+{% include "../../../addons/flux-cronjob-deployment/README.md" %}
+
+{% include "template/usage.md" %}
+
