+++ docs/templates_/addons/flux-service-deployment.md
@@ -1,0 +1,5 @@
+
+{% include "../../../addons/flux-service-deployment/README.md" %}
+
+{% include "template/usage.md" %}
+
