+++ docs/templates_/addons/github-workflows.md
@@ -1,0 +1,5 @@
+
+{% include "../../../addons/github-workflows/README.md" %}
+
+{% include "template/usage.md" %}
+
