+++ docs/templates_/addons/mkdocs.md
@@ -1,0 +1,5 @@
+
+{% include "../../../addons/mkdocs/README.md" %}
+
+{% include "template/usage.md" %}
+
