+++ docs/templates_/addons/ness-lambda.md
@@ -1,0 +1,5 @@
+
+{% include "../../../addons/ness-lambda/README.md" %}
+
+{% include "template/usage.md" %}
+
