+++ docs/templates_/bases/sfmc-app.md
@@ -1,0 +1,5 @@
+
+{% include "../../../bases/sfmc-app/README.md" %}
+
+{% include "template/usage.md" %}
+
