+++ docs/templates_/bases/simple-app.md
@@ -1,0 +1,6 @@
+---
+tags: [simple-app, template]
+---
+{% include "../../../bases/simple-app/README.md" %}
+
+{% include "template/usage.md" %}
