--- docs/templates_/discovery-app-layer.md
@@ -1,1 +1,0 @@
-{% include "../../addons/discovery-app-layer/README.md" %}
