+++ docs/templates_/external/template-infra.md
@@ -1,0 +1,5 @@
+# Platform Infrastructure Template
+
+A template for creating team infrastructure repositories.
+
+[https://github.com/tx-pts-dai/template-infra]()
