--- docs/templates_/github-workflows.md
@@ -1,1 +1,0 @@
-{% include "../../addons/github-workflows/README.md" %}
