+++ docs/templates_/index.md
@@ -1,1 +1,13 @@
 # Templates
+
+# Bases
+
+Base templates are designed to be used as a bootstrap template for new projects. This are opinionated the golden paths for the Tamedia Core engineering organisation.
+
+# Addons
+
+Addon templates are designed to be used with already existing projects. They are opinionated and provide a structured approach to adding new features to your project.
+
+# External
+
+Templates stored in external repositories.
