--- docs/templates_/mkdocs.md
@@ -1,1 +1,0 @@
-{% include "../../addons/mkdocs/README.md" %}
