--- docs/templates_/ness-lambda.md
@@ -1,1 +1,0 @@
-{% include "../../addons/ness-lambda/README.md" %}
